plumb through purego that removes pkcs11 support
